Remove an Unpin impl for JoinHandle.
We should just let the compiler determine it. We wrap over several JoinHandles:
tokio, async-std, futures RemoteHandle. Tokio is the only one that adds a similar
Unpin impl.

I have a hard time figuring out whether it's correct. We don't necessarily
hold a T, but we do hold a future that produces a T, so which might hold
one across an await point.